fix(runner): parse v3 extension varId as unsigned byte (k.toInt.toByte)
SpecExtract encodes the varId unsigned (id & 0xff, so a negative Byte becomes 128-255) and the schema permits 0-255, but Runner/EvalConformanceTest parsed it with k.toByte (signed -128..127) -> NumberFormatException on varId >= 128. k.toInt.toByte is the correct inverse. Latent (current vectors use varId 11); caught by the pre-merge review. jvm 72/72 green.

+++ jvm-blesser/src/main/scala/santa/runner/Runner.scala
@@ -36,7 +36,7 @@ object Runner {
         val inputs = c.downField("inputs").values.getOrElse(Vector.empty).toVector.map { inp =>
           val ext = inp.hcursor.downField("extension")
           ext.keys.getOrElse(Iterable.empty).iterator.map { k =>
-            k.toByte -> EvalCore.decodeInputConstant(
+            k.toInt.toByte -> EvalCore.decodeInputConstant(
               ext.downField(k).focus.getOrElse(sys.error(s"v3 entry '$name': missing extension value for key $k")))
           }.toMap
         }
